fix: address review comments on PR #16
